Create a `cached_state` module
Move the test utility functions related to using a cached Zebra state
into the module.
