+++ zebrad/tests/common/mod.rs
@@ -9,6 +9,7 @@
 //! to avoid compiling an empty "common" test binary:
 //! https://doc.rust-lang.org/book/ch11-03-test-organization.html#submodules-in-integration-tests
 
+pub mod cached_state;
 pub mod check;
 pub mod config;
 pub mod launch;
